requirements.txt
Django==3.2
Pillow==9.3.0
djangorestframework==3.14.0
django-filter==22.1
djangorestframework-jwt==1.11.0
markdown==3.4.1
django-cors-headers==3.13.0
